Refine shift inspect CLI
