+++ crates/shift-cli/Cargo.toml
@@ -2,13 +2,14 @@
 name = "shift-cli"
 version = "0.1.0"
 edition = "2024"
+description = "Read-only command-line tools for Shift source packages"
+readme = "README.md"
 
 [[bin]]
 name = "shift"
 path = "src/main.rs"
 
 [dependencies]
-anstream = "0.6"
 anstyle = "1"
 clap = { version = "4.5.40", features = ["derive"] }
 comfy-table = "7.2"
